Changed wakeup_time to int64 and changed how interrupts are disabled and enabled in timer.c

+++ pintos/src/threads/thread.h
@@ -106,7 +106,7 @@ struct thread
     /* Owned by thread.c. */
     unsigned magic;                     /* Detects stack overflow. */
     // Added for sleep function (Jim)
-    int wakeup_time;
+    int64_t wakeup_time;
   };
 
 /* If false (default), use round-robin scheduler.
